fix: recursive call to getDevices when the browser does not support audiooutput like Safari and Firefox.
The issue was happening when the browser does not support audiooutput and
it returns device list with empty array when the code received empty device list it was trying to call getUserMedia and getDevices to fullfiled with data,
but as the browser does not support it was executing the getDevices method in infinite loop.

To solve it we will only execute getDevices in recursion if the device list has data with label and deviceid empty.
This means that the browser supports audiooutput but the user does not allowed the permission to access devices yet.